Gears test

+++ cs296_base_code/src/dominos.hpp
@@ -40,7 +40,7 @@ namespace cs296
       return new dominos_t;
     }
 
-    b2Body* createGear(int, uint16, float, float);
+    b2Body* createGear(float, uint16, float, float);
   };
 }
